Health monitoring system controlled by IOT
The Internet of Things (IOT) is a system of networked objects that can communicate and
gather data thanks to software, sensors, network connectivity, and electronic devices that
allows the system to easily exchange the data. This work aims to collect the reading of
essential health parameters like body temperature, blood oxygen, and pulse rate by using
different biosensors. The complete system is on the Arduino UNO and NodeMCU. The
system collects the various inputs provided by the sensors which work with the voice
command and it sends the data to the concerned authorities by using the
Thingspeak platform and displays the data to the user on the LCD, the proposed system can
be very useful for the nursing staff and doctors to monitor the health status of the individual.
So, that better treatment can be provided to the patient. This can also help in improving
health monitoring in rural parts of India.